chore(core): renamed some files and functions to be more idiomatic and rust complaint
