+++ parser/rust/analyzer/libs/parsers/src/lib.rs
@@ -6,8 +6,8 @@ pub mod socket;
 pub mod network;
 pub mod listen;
 pub mod fcntl;
-pub mod readwrite;
-pub mod recvsend;
+pub mod read_write;
+pub mod recv_send;
 pub mod clone;
 pub mod close;
 pub mod munmap;
