+++ parser/rust/analyzer/libs/parsers/src/read_write.rs
@@ -4,7 +4,7 @@ use wrappers::trackers::Trackable;
 use serde::{Deserialize, Serialize};
 use std::any::Any;
 use std::rc::Rc;
-use trackers::descriptors::Descs;
+use trackers::fd_table::Descs;
 
 #[derive(Debug, Serialize,Deserialize)]
 pub struct ReadWriteArgs {
@@ -88,7 +88,7 @@ impl Trackable for ReadWriteTrack {
         }
         
 
-        let uuid = match descs.get_by_descriptor_number(args.fd) {
+        let uuid = match descs.get_fd(args.fd) {
             Some(record) => &record.uuid,
             None => {
                 return Err("No uuid found".to_string())
