+++ parser/rust/analyzer/libs/trackers/src/lib.rs
@@ -1,1 +1,1 @@
-pub mod descriptors;
+pub mod fd_table;
